# Conflicts:
#	hobbiton.iml
#	src/main/java/com/hackathon/hobbiton/controller/authorization/Authorization.java

+++ src/main/java/com/hackathon/hobbiton/database/mapper/UserMapper.java
@@ -12,9 +12,9 @@ public class UserMapper {
         user.setLogin(rs.getString("login"));
         user.setEmail(rs.getString("email"));
         user.setGender(rs.getString("gender"));
-        user.setPoints(rs.getLong("points"));
-        user.setSubscription(rs.getLong("subscription"));
-        user.setFollowers(rs.getLong("followers"));
+        user.setPoints(rs.getInt("points"));
+        user.setSubscription(rs.getInt("subscription"));
+        user.setFollowers(rs.getInt("followers"));
         return user;
     }
 }
